Bug fixes and optimization
Fix that when `Tokenizer` and `Parser` tries to report an `Error`,
the program crash. This is because file information is not given
to these two classes.

Fix the format problem of error `WRONG_IF_CONDITION`. The raw "repr"
of the got type (`<...BuiltinIntType object at ...>`) is printed.

`acaciamc/mccmdgen/generator.py` uses context manager to manage
`If`, `InterfaceDef` and `FuncDef` now.