A lot of changes on `nonetype`
`nonetype` is now an "unstorable" type, which means it is not stored
in Minecraft, and you cannot create variables of it (Just like strings).
So, why? Suppose that `nonetype` is "storable".
Then this is legal:
```
import print
x = print.tell("Hello, world!")
x = print.tell("The second message")
```
Commands *are* generated on line 2 & 3 and `x` can even be re-assigned!

Add keyword `None` referring to a `nonetype` object.
(PS. Suppose `nonetype` is "storable", `None` can even be assigned!)