Changes on symbol management
Contents added to symbol table used to be immutable.
For instance, these are not allowed:
```
a -> 1
a -> 2 # `BIND_TARGET_EXISTS`
```
```
a = 10
a = 20 # OK, since assignment does not change symbol table
import a # `MODULE_NAME_CONFLICT`
```
Now these are OK. However, it's still not recommended to
change the type of an identifier.

Builtin contents is not added by setting `outer` of scopes to a
"basic_scope" with all the builtin names, but by setting the `builtins`
of the scopes.