Update `powc` function to `pow` in module `math`
`math.powc` is used to do the power operation before and it only
accepts integer literal as the second argument.
Now we don't have this limit. An internal module called `_math` is
written in Acacia and use a while statement to do power calculations.

A little change on interface of compiler and AST so that
the new `math.pow` can be implemented easily.