Change on tokenizer (mainly about indentations)
Now any number of spaces for indentation is accepted even without "-i"
compile option. Different number of spaces can even mix together as long
as statements on same level have the same number of spaces indented.
Thus, "-i" is removed.

Error `WRONG_INDENT` is removed since there is no fixed number of spaces
for indentations. Now `EMPTY_BLOCK` is always raised when an indented
block is expected but didn't present.

New error `INVALID_DEDENT` is added. This is raised when the number of
spaces for unindentation does not match any level.

New error `EOF_AFTER_CONTINUATION` is added. This is raised when there
is no new line after line continuation backslash.

Since the tokenizer gets a huge update, there are some tiny edge cases
where old tokenizer and new one produce different results.
For example:
    if x:
        #* yyy
    *# x += 1
works for new tokenizer because it treats line 2 and 3 as a whole line
and "x += 1" is considered indented and is in body of if statement. Old
tokenizer raises a `WRONG_INDENT` error on it.

Tokenizer now requires a file-like object that supports `readline` as
source file argument, rather than `str`.