Remove `iterlib.map` and change on iterators
Iterators will no longer be able to execute commands now. `iterlib.map`
relies on that and it is now deleted. This is bacause `map`, as a
function, should run the given function in its body, rather than
executing commands in the for-in structure where the return value of
`map` is used.