`array` and `map` types
Add 2 compile time only data types `array` and `map` which are similar
to Python `list` and `dict`, respectively.

`array` and `map` can be created using curly brackets "{}". `{1, 2}`
creates an `array` with elements 1 and 2. `{}` creates an empty `array`.
`{:}` creates an empty map. `{1: "a", 2: "b"}` creates a mapping that
maps 1 to "a" and 2 to "b".

New protocol `map_hash` is added to allow an expression become key of a
`map`. So far `map` is not working because no class implements it.