Fix inverted `and` as operand of another `and`
When an `and` group is `not`ed, and it is used as an operand of another
`and` group, the evaluation of the expression will probably be wrong.
Since `or`'s mechanism is to convert into `not` and `and` combination,
using a regular (un-inverted) `and` group as an operand of an `or` group
will also result in wrong generated commands.
An example to reproduce:
    y = True
    x = 11
    boolean = (y and (x > 10)) or (x < 0)
    if boolean:
        /say Should be printed, but actually not.