Optimize: fix a potential bug in symbol.py
First, `super()` with no argument must be called just in method level,
not in an embedded function. Secondly, an embedded function in not
necessary here. Thirdly, yielding directly from a `super` object seems
not to work even if the base class defines `__iter__`.
`ScopedSymbolTable.__iter__` is not used so far, so the problem is
potential.