Declaration update
Rationale
=====================
we separate declaration and assignment in this update! Acacia
is a strongly typed language and I hope to enphasize this through this
update. Another reason is that `Engroup.t` design is quite strange, I
hope to eliminate it through this change.

Syntax change
=====================
The original assignment statement using `=` will no longer be able to
create a new variable. The variable declaration statement is introduced
to explicitly declare a variable:
    x: int
This defines a variable of integer type with unknown value. Then we may
assign to it:
    x = 10
Initial value may be given in a declaration:
    x: int = 0
Or, even shorter:
    x := 0
`:=` declares a variable with an initial value, of which type is
inferred by the compiler from the initial value.

Some data types may accept arguments to initialize the value:
    struct T:
        x: int
    t: T | (x=10)
This is done by adding `|` and listing arguments just like calling
a function.

Other major changes
=====================
- Remove `Engroup.all_entities`. Use `v: Engroup | (all_entities=True)`
- Remove `Engroup.__new__`.
- Remove `Engroup.t`.
- `e: entity` does not summon an entity. It creates a reference to an
  unknown entity. Still use `e := Object()`.
- Due to these backward-incompatible changes, we update all tests,
  demos, READMEs and standard library module `_math`.

Minor changes
=====================
- Internal changes:
  * `AcaciaExpr.new_var`
  * `export_need_tmp`
  * Newly summoned entities are now `EntityReference`s returned by call
    to entity templates
- Change on test/test_ast.py so that visualized `ArgumentTable`s won't
  have special representation.