Merge `for entity` statement into for-in statement
This code:
    for entity x in y:
        do_something(x)
Used to iterate over every `entity` in entity group `y` and bound them
to `x` one by one and run function `do_something` with each `x` as
argument.
Now the keyword `entity` should be deleted and use regular for-in syntax
instead. Note that though the syntax now looks the same as compile-time
loop for-in on a `map` or an `array`, they are handled differently. I
hope this would result in better readability.