Virtual method and fix on entity system
Add "virtual" keyword which declares a virtual method and works very
similar to virtual methods in C++. Before this update, all methods are
actually implicitly "virtual". Virtual methods are entity methods whose
behavior can be overridden in derived classes. As opposed to non-virtual
methods, the overriding behavior is preserved even if there is no
compile-time information about the actual template of an entity. That is
to say, if a derived entity is handled using a function that accepts an
entity of the base template as argument, a call to its virtual method
would invoke the behavior defined in the derived template.
    entity X:
        virtual def f():
            /say f in X
        def b():
            /say b in X
    entity Y extends X:
        def f():
            /say f in Y
        def b():
            /say b in Y
    def f(x: entity(X)):
        x.f()
        x.b()
        X@x.f()  # tell compiler to use f in X
    x := X()
    y := Y()
    f(x)  # f in X; b in X; f in X
    f(y)  # f in Y; b in X; f in X

Current overridden methods should be compatible with the "virtual"
definition. For example, this is not allowed now:
    entity X:
        virtual def foo():
            pass
    entity Y extends X:
        def foo(i: int):
            pass
    X().foo()  # error
Before this update the implementation in Y got hidden implicitly for
this `X().foo()` call.

Fix 2 bugs about entity casting.
    1. An entity that is casted does not select the original entity,
    instead it selects nothing.
    2. Calling a method of a casted entity will fail to pass "self"
    environment.
    (Actually 2 huge bugs...)
To reproduce:
    import world
    import print

    entity X:
        x: int

        def f():
            self.x = 1

    x := X()
    x.x = 0
    X@x.f()
    print.tell(print.format(
        "Bug 1: should be 1 but is actually %0", x.x
    ))
    c := X@x  # `c` is selecting no entity ...
    world.kill(c)
    world.msg_say(x, "Bug 2: should be killed but I'm alive")