Fix a inline function bug about context
The inline function should be executed under the context (i.e. scope,
result value, self value) where it is defined, not the context where it
is called.

Add this code to mod.aca test file that could reproduce this bug:
    # mod.aca
    inline def get_inline_func():
        a: int = 0
        inline def inline_func():
            a += 1
            print.tell(print.format("a: %0", a))
        result -> inline_func
    # module.aca
    from .mod import get_inline_func
    f -> get_inline_func()
    f()  # NAME_NOT_DEFINED
    f()

Inline entity methods are now called without "execute as ..." context,
instead the related entity is directly used in generated commands.