Change on filter in `filter` & `drop` of `Engroup`
Change about use of filter in `filter` and `drop` method of `Engroup`:
    Old: select entities that satisfies the filter from all entities in
    the world.
    New: select entities that satisfies the filter from all entities
    that are in the group the world.
This does make a difference when using methods like `random` of
`Enfilter`. When seeing this:
    g: Engroup
    # ... (add something to `g`)
    g.drop(Enfilter().random("armor_stand", limit=1))
I think about "dropping 1 random armor_stand from `g`", not "selecting
1 random armor_stand from all the armor_stands in the world and if it is
in `g` we drop it".
`world.is_entity` and `select` of `Engroup` still selects entities from
all the entities in the world.