Module `world`: strings as `Block` and `Item`
Now the functions provided by `world` support using strings to replace
the original `Block` and `Item` object if it is a simple block or item.
Instead of:
    import world
    world.setblock(Pos(0, 0, 0), world.Block("diamond_block"))
    world.replaceitem_block(
        Pos(0, 0, 0), slot=0, item=world.Item("diamond")
    )
We have:
    import world
    world.setblock(Pos(0, 0, 0), "diamond_block")
    world.replaceitem_block(Pos(0, 0, 0), slot=0, item="diamond")
`Block` and `Item` are still useful for blocks with states or items with
components or data value.

Shorten code in READMEs and brief.aca using this feature.