Improve function inliner in optimizer
Now functions called with execute environment that consists of only
"if" or "unless" (any subsubcommand) can be inlined. Before this update
only "if" or "unless" with "score" subsubcommand is allowed.

In function calls that satisfy the inline condition, if callee only has
1 command in it, tmp is not used anyway.

Refactor inliner's tmp check a little bit to make it more efficient.