Add constructor system
Constructor functions are a special type of function that when it gets
called directly on the right of `=` assignment, special conversion will
be made to avoid creation of temporary variables.
    entity X:
        i: int
    e1: Entity
    e1 = X()  # calls "reconstructor" of `X`
    e2: Entity = X()  # works same as e1
    e3 := X()  # becomes `e3 -> X()`
    i1 := X().i  # unchanged
    e4 -> X()  # unchanged
This results in less commands when assigning newly created variables.

Now you can initialize a struct by calling struct template.
Instead of:
    x: SomeStruct | (some_field=10)
You can do:
    x := SomeStruct(some_field=10)