Formatted string literal
Now in string you can use ${formatted expr} just like in commands. `\$`
escape also works and result in a plain `$` character. The strings may
be nested:
    /command ${lv1 + ${"lv2 first"} + ${"lv2 second ${3}"}}
The expressions are converted to string in the same way as they are in
commands. If it is not convertable, `INVALID_FEXPR` is raised.

Rename `INVALID_CMD_FORMATTING` -> `INVALID_FEXPR`