Improve use of tmp value in "if" and "while"
Add an ABC `SupportsAsExecute` which is a boolean expression that can
trigger optimizations in "if" (without "else" or "elif") and "while"
statements.

Fix integer comparison expressions in "and" group (or equivelant things)
that has ">= int.MIN" or "<= int.MAX". Steps to reproduce:
    x := 1
    y := int.MIN <= x <= int.MAX
    z := x <= 2147483647 and y
Now `y` and `z` are True instead of generating a command with syntax
error and leading to an unexpected behavior.

Refactor `AndGroup` a bit and rename it to `WildBool`.