Fix calling the same function in one expression
When calling the same non-inline function in one same expression the
result values may be wrong. To reproduce:
    def foo(x: int) -> int:
        result = x + 10
    v := foo(1) + foo(2)
    # `v` is 24 now...
Fixing this *reduces* the performance of calling a non-inline function
because we have to make a copy of *every* result... But it's hard to fix
it in another way. There might be optimizations for this in the future.