Assignment target can now have any expression AST
No more restrictions about AST type on assignment target now! Before we
can only assign to several particular ASTs, now it can be any
expression type, but still there will be a check during command
generation to see if the value is really assignable semantically. So now
we are able to do:
    x := 0
    inline def foo() -> int:
        x += 1
        result -> x
    foo() *= 2
instead of:
    tmp -> foo()
    tmp *= 2