Fix bugs related to instance creation
Fix that when unable to instantiate in compile time functions, the type
shown in error message is always "type".

Fix that when __new__ is only available in either of runtime or compile
time, unexpected error occurs when calling in the one that it does not
support. Now a CANT_CREATE_INSTANCE error is raised. A way to reproduce
is to call `schedule.Task` inside a compile time function.