Support assignment and entity field for `None`
It would not make sense to make a variable declared using ":" syntax
unassignable. Consider the following:
    x := None  # ok
    x: None = None  # compilation error
Now both are OK and all variables of `None` type can be re-assigned a
value. However, assignment to the `None` keyword is not allowed.

Variables of `None` type may be entity field now. This may be useful
when doing something similar to a generic, for example:
    inline def pair(const T) -> const Any:
        entity Pair:
            first: T
            second: T
        result Pair
    pair(int)()  # entity that holds 2 `int`s.
    pair(None)()  # entity that holds no data.
    # (This is not quite a proper example, by the way :/)