New `new` system to alter entity spawning
The old `SpawnInfo` system has some disadvantages:
    - `__spawn__` is called at compile time and so the spawn info must
      be determined at compile time, meaning that we cannot alter
      the info at runtime.
    - `__init__` and `__spawn__` both receive the arguments passed into
      the entity template call, but some arguments are used only by one
      of them, leaving an unused argument in the other function.
    - If a subtemplate changes the parameter of either `__init__` or
      `__spawn__`, then both need to be overridden.
    - `static inline def __spawn__(...) -> const SpawnInfo` is very
      looooong.
Therefore, a new system is added, which makes use of a new keyword
named `new`. `__spawn__` and `__init__` are now combined into a single
`new` method. In (and only in) a `new` method it is possible to call
other `new` methods to spawn the entity through the syntax `X.new(...)`
or one may just summon the entity using a direct `new(...)` call.
`new(...)` receives same arguments as the old `SpawnInfo`, and it really
spawns the entity and binds it to `self`, meaning that if the entity is
not yet spawned in a `new` method, then value of `self` is undefined.
`X.new(...)` calls the `new` method of template `X`. `new(...)` is in
fact same as `Entity.new(...)`.

The entity template id is now implemented using scoreboard instead of
tags for easier implementation of the `new` constructor.