Fix and improve `ExternEngroup` and ctor system
Now `ExternEngroup` is no longer a generic because the type of an
`ExternEngroup` has nothing to do with the template it is going to be
resolved as. Now the template needs to be passed to `resolve`.

The `ConstructorFunction` system is used to optimize commands generated.
It is designed to work in assignment or result statements but a mistake
was made and led to the fact that only `:=` statements make use of this
system. Now they should work as expected.