Better interface path syntax
Now the interface path matches its representation in Minecraft: slashes
are used for subdirectories and characters other than letters, numbers
and underscore are allowed (except parens). The path is now a whole
token so no physical line breaks can be made in between the path. An
alternative quote syntax is added for function names with parens or
interpolation. Examples:
    interface this_still_works:
        pass
    interface foo/bar:
        pass
    interface dots.here:
        pass
    interface "alternative/syntax":
        pass
    const string = "foo"
    interface "formatted/${string}":
        pass