Initial commit of V3 refactor
Yep, this commit deleted almost every single file in the project. This
is going to be the biggest refactor that has ever happened on this
project.

I only kept the tokenizer and the parser parts. (And a lot of changes
have been made on those two parts in this commit as well.) The reason is
that I believe having a `Generator` to handle everything is too complex.
Modern compilers do a LOT of passes, and each pass only focuses on a
small portion of the big compiling task. Current design only makes 3
passes -- tokenizer, parser, and then code generator. The code generator
is doing way too much. The new design will have multiple passes, and
a custom IR will probably be established to make the compiling process
even "slower" (but probably smoother).

Another big change is about general coding workflow. I am adopting unit
tests. I will try to write more comments, but reduce useless comments.
The type checker is now on, so all those type hints are not just a hint
to programmers but also a way to discover bugs.

And for what we have done specifically in this commit:
    * The diagnostic system is better. Now we use fancy ANSI sequence
      to display colored text. The portion of code that triggers the
      error is displayed and highlighted. We do not rely on Python's
      builtin `str.format()` now, instead we can add custom arguments to
      diagnostic messages. Source locations can now be ranges (like
      foo.aca:10:5-9).
    * The structure of a few AST nodes were changed. Not every AST node
      has source location now.
    * Fix crash when the file ends with an identifier or an incomplete
      unicode escape (\x, \u or \U). In fact, the type checker found
      these two bugs.